+++ crates/miden-protocol/src/address/routing_parameters.rs
@@ -107,8 +107,8 @@ impl RoutingParameters {
 
     /// Returns the note tag length preference.
     ///
-    /// This is guaranteed to be in range `0..=32` (e.g. the maximum of
-    /// [`NoteTag::MAX_ACCOUNT_TARGET_TAG_LENGTH`]).
+    /// This is guaranteed to be in range `0..=32` (i.e. at most
+    /// [`NoteTag::MAX_ACCOUNT_TARGET_TAG_LENGTH `]).
     pub fn note_tag_len(&self) -> Option<u8> {
         self.note_tag_len
     }